add authorization checks

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const path = require("path");
 
 const { validationResult } = require("express-validator");
 const Post = require("../models/post");
+const User = require('../models/user');
 
 const fakeCreator = {
   name: "Mileena",
@@ -70,20 +71,29 @@ exports.createPost = (req, res, next) => {
 
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
 
   const post = new Post({
     title: title,
     content: content,
-    creator: fakeCreator,
+    creator: req.userId,
     imageURL: imageURL,
   });
   post
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result=>{
       res.status(201).json({
         message: "Post created succesfully!",
-        post: result,
+        post: post,
+        creator:{_id: creator._id, name: creator.name }
       });
     })
     .catch((err) => {
@@ -118,6 +128,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      
+      if (post.creator.toString() !== req.userId.toString()){
+        const error = new Error('Not authorized');
+        error.statusCode = 404;
+        throw error;
+      }
+
       if (imageURL !== post.imageURL) {
         clearImage(post.imageURL);
       }
@@ -151,11 +168,24 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      //check logged in user
+      
+      if (post.creator.toString() !== req.userId.toString()){
+        const error = new Error('Not authorized');
+        error.statusCode = 404;
+        throw error;
+      }
+      
       clearImage(post.imageURL);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then(user =>{
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Post deleted" });
     })
     .catch((err) => {
